expose client apis
