+++ src/index.ts
@@ -1,11 +1,15 @@
 import * as trpcExpressAdapter from '@/trpc/server/handler';
 import { WebSocket } from 'ws';
-import * as clientApi from '@/trpc/client/api';
-import * as clientConnectionsApi from '@/trpc/client/connections';
+import * as api from '@/trpc/client/api';
+import * as usersApi from '@/trpc/client/users';
+import * as messagesApi from '@/trpc/client/messages';
+import * as connectionsApi from '@/trpc/client/connections';
 
 export {
     trpcExpressAdapter,
     WebSocket,
-    clientApi,
-    clientConnectionsApi
+    api,
+    usersApi,
+    messagesApi,
+    connectionsApi
 };
